Move AppManager into android.internal package.
