+++ formula-android/src/main/java/com/instacart/formula/android/FragmentEnvironment.kt
@@ -1,7 +1,5 @@
 package com.instacart.formula.android
 
-import com.instacart.formula.android.FragmentKey
-
 data class FragmentEnvironment(
     val onScreenError: (FragmentKey, Throwable) -> Unit = { _, it -> throw it }
 )
