+++ formula-android/src/main/java/com/instacart/formula/android/internal/AppManager.kt
@@ -1,12 +1,10 @@
-package com.instacart.formula.android
+package com.instacart.formula.android.internal
 
 import android.app.Activity
 import android.app.Application
 import android.os.Bundle
 import androidx.fragment.app.FragmentActivity
-import com.instacart.formula.android.internal.ActivityManager
 import com.instacart.formula.android.events.ActivityResult
-import com.instacart.formula.android.internal.ActivityStoreFactory
 import java.util.UUID
 
 /**
